requirements.txt
streamlit
requests
langdetect
deep-translator
pdfminer.six
spacy